+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/util/PaddingToolTest.kt
@@ -5,19 +5,15 @@ import de.rki.coronawarnapp.server.protocols.internal.v2.PresenceTracingParamete
 .PresenceTracingPlausibleDeniabilityParameters.NumberOfFakeCheckInsFunctionParametersOrBuilder
 import de.rki.coronawarnapp.server.protocols.internal.v2.PresenceTracingParametersOuterClass
 .PresenceTracingPlausibleDeniabilityParameters.NumberOfFakeCheckInsFunctionParameters
-import de.rki.coronawarnapp.server.protocols.internal.v2.RiskCalculationParametersOuterClass.Range
 import de.rki.coronawarnapp.util.PaddingTool.determineNumberOfFakeCheckIns
 import de.rki.coronawarnapp.util.PaddingTool.equation
+import io.kotest.matchers.doubles.shouldBeGreaterThan
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.extension.ExtensionContext
 import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.Arguments
-import org.junit.jupiter.params.provider.ArgumentsProvider
 import org.junit.jupiter.params.provider.ArgumentsSource
 import testhelpers.BaseTest
 import timber.log.Timber
-import java.util.stream.Stream
 import kotlin.math.abs
 import kotlin.random.Random
 
@@ -88,126 +84,22 @@ class PaddingToolTest : BaseTest() {
     }
 
     @ParameterizedTest
-    @ArgumentsSource(NumberOfFakeCheckInsProvider::class)
+    @ArgumentsSource(ZeroFakeCheckInsProvider::class)
     fun `Determine Number Of Fake CheckIns = 0`(
-        inputParam: PlausibleDeniabilityParametersContainer,
-        inputNumberOfCheckIns: Int,
-        expectedOutput: Double
+        plausibleDeniabilityParameters: PlausibleDeniabilityParametersContainer,
+        numberOfCheckIns: Int,
+        expected: Double
     ) {
-        inputParam.determineNumberOfFakeCheckIns(inputNumberOfCheckIns) shouldBe expectedOutput
+        plausibleDeniabilityParameters.determineNumberOfFakeCheckIns(numberOfCheckIns) shouldBe expected
     }
-}
 
-/*
-  "data": [
-        { "x": -1, "fx": 26.0308204914619 },
-        { "x": 0, "fx": 61.59927895860675 },
-        { "x": 1, "fx": 94.75879170760312 }
-      ]
- */
-class EquationProvider : ArgumentsProvider {
-    override fun provideArguments(context: ExtensionContext?): Stream<out Arguments> =
-        Stream.of(
-            Arguments.of(-1.0, 26.03082049146190),
-            Arguments.of(0.0, 61.59927895860675),
-            Arguments.of(1.0, 94.75879170760312)
-        )
-}
-
-class NumberOfFakeCheckInsProvider : ArgumentsProvider {
-    override fun provideArguments(context: ExtensionContext?): Stream<out Arguments> =
-        Stream.of(
-            /*
-              {
-              "description": "returns 0 if there are no check-ins and the respective probability is 0",
-              "parameters": {
-                "numberOfCheckInsInDatabaseTable": 0,
-                "appConfigParameters": {
-                  "probabilityToFakeCheckInsIfSomeCheckIns": 1,
-                  "probabilityToFakeCheckInsIfNoCheckIns": 0,
-                  "numberOfFakeCheckInsFunctionParameters": [
-                    {
-                      "randomNumberRange": { "min": -9999, "max": 9999 },
-                      "p": 100,
-                      "q": 1.4,
-                      "r": -1,
-                      "s": 0.8,
-                      "t": -1.5,
-                      "u": 2,
-                      "a": 0,
-                      "b": 0,
-                      "c": 0
-                    }
-                  ]
-                }
-              },
-              "expNumberOfCheckIns": 0
-            }
-           */
-            Arguments.of(
-                PlausibleDeniabilityParametersContainer(
-                    probabilityToFakeCheckInsIfSomeCheckIns = 1.0,
-                    probabilityToFakeCheckInsIfNoCheckIns = 0.0,
-                    numberOfFakeCheckInsFunctionParameters = functionParamsList()
-                ),
-                0, // Number of CheckIns in Database
-                0 // Expected number of fake CheckIns
-            ),
-
-            /*
-                {
-                  "description": "returns 0 if there are some check-ins and the respective probability is 0",
-                  "parameters": {
-                    "numberOfCheckInsInDatabaseTable": 10,
-                    "appConfigParameters": {
-                      "probabilityToFakeCheckInsIfSomeCheckIns": 0,
-                      "probabilityToFakeCheckInsIfNoCheckIns": 1,
-                      "numberOfFakeCheckInsFunctionParameters": [
-                        {
-                          "randomNumberRange": { "min": -9999, "max": 9999 },
-                          "p": 100,
-                          "q": 1.4,
-                          "r": -1,
-                          "s": 0.8,
-                          "t": -1.5,
-                          "u": 2,
-                          "a": 0,
-                          "b": 0,
-                          "c": 0
-                        }
-                      ]
-                    }
-                  },
-                  "expNumberOfCheckIns": 0
-                }
-             */
-            Arguments.of(
-                PlausibleDeniabilityParametersContainer(
-                    probabilityToFakeCheckInsIfSomeCheckIns = 0.0,
-                    probabilityToFakeCheckInsIfNoCheckIns = 1.0,
-                    numberOfFakeCheckInsFunctionParameters = functionParamsList()
-                ),
-                10, // Number of CheckIns in Database
-                0 // Expected number of fake CheckIns
-            )
-        )
-
-    private fun functionParamsList() = listOf(
-        NumberOfFakeCheckInsFunctionParameters.newBuilder()
-            .setRandomNumberRange(
-                Range.newBuilder()
-                    .setMin(-9999.0)
-                    .setMax(9999.0)
-            )
-            .setP(100.0)
-            .setQ(1.4)
-            .setR(-1.0)
-            .setS(0.8)
-            .setT(-1.5)
-            .setU(2.0)
-            .setA(0.0)
-            .setB(0.0)
-            .setC(0.0)
-            .build()
-    )
+    @ParameterizedTest
+    @ArgumentsSource(MoreThanZeroFakeCheckInsProvider::class)
+    fun `Determine Number Of Fake CheckIns more than 0`(
+        plausibleDeniabilityParameters: PlausibleDeniabilityParametersContainer,
+        numberOfCheckIns: Int,
+        expected: Double
+    ) {
+        plausibleDeniabilityParameters.determineNumberOfFakeCheckIns(numberOfCheckIns) shouldBeGreaterThan expected
+    }
 }
